+++ microcontroller-2/hello-arm/lm3s9b90/dump.c
@@ -70,8 +70,8 @@ void dump(char* s){
 	char tstr[9];
 	if(s){
 		s=strstrip(s);
-		if(isalpha(*s)){
-			while(isalpha(*s))
+		if(isalpha((uint8_t)(*s))){
+			while(isalpha((uint8_t)(*s)))
 				++s;
 		}
 		char* eptr;
